agent: Extract cloudwatch specific functions
